+++ tf2_kdl/docs/mainpage.dox
@@ -11,9 +11,9 @@ by the Orocos KDL (Kinematics and Dynamics Library) API (see http://www.orocos.o
 See the <A HREF="http://wiki.ros.org/tf2/Tutorials/Migration/DataConversions">Conversions overview</A>
 wiki page for more information about datatype conversion in tf2.
 
-\section codeapi Code API 
+\section codeapi Code API
 
-This library consists of one header only, tf2_kdl/tf2_kdl.h, which consists mostly of
+This library consists of one header only, tf2_kdl/tf2_kdl.hpp, which consists mostly of
 specializations of template functions defined in tf2/convert.h.
 
 */
